pyproject.toml
[project]
name = "irods_prefect"
version = "0.1.0"
description = "Prefect v3 workflows for iRODS interactions"
readme = "README.md"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.10"
dependencies = [
    "prefect>=3.3.5",
    "python-irodsclient>=3.1.0",
    "fastapi>=0.115.12",
    "uvicorn>=0.34.1",
    "pydantic>=2.11.3",
    "pydantic_settings>=2.9.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]


[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"


[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.1.0",
    "black>=23.9.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "flake8>=6.1.0",
]

[tool.setuptools]
packages = ["irods_prefect"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
python_classes = "Test*"
